requirements.txt
astroid==1.6.3
attrs==17.4.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
coverage==4.5.1
coveralls==1.3.0
docopt==0.6.2
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.1
mccabe==0.6.1
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==4.1.0
numpy==1.14.2
pandas==0.22.0
Pillow==5.1.0
pkginfo==1.4.2
pluggy==0.6.0
py==1.5.3
pylint==1.8.4
pytest==3.5.1
pytest-cov==2.5.1
python-dateutil==2.7.2
python-pptx==0.6.8
pytz==2018.4
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
tqdm==4.23.0
twine==1.11.0
urllib3==1.22
wincertstore==0.2
wrapt==1.10.11
XlsxWriter==1.0.4
